requirements.txt
streamlit>=1.48.0
pandas>=1.4.0
openpyxl>=3.1.0
gspread>=5.0.0
google-auth>=2.0.0